Реализовал resourced-based авторизацию для существующих методов

+++ DeliveryDeck_Backend_Final/Controllers/CartController.cs
@@ -1,8 +1,7 @@
-﻿using DeliveryDeck_Backend_Final.ClaimAuthorize;
-using DeliveryDeck_Backend_Final.Common.CustomPermissions;
+﻿using DeliveryDeck_Backend_Final.Common.CustomPermissions;
 using DeliveryDeck_Backend_Final.Common.DTO.Backend;
 using DeliveryDeck_Backend_Final.Common.Interfaces.Backend;
-using DeliveryDeck_Backend_Final.Common.Utils;
+using DeliveryDeck_Backend_Final.Filters;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.ModelBinding;
@@ -15,9 +14,11 @@ namespace DeliveryDeck_Backend_Final.Controllers
     public class CartController : AuthorizeController
     {
         private readonly ICartService _cartService;
-        public CartController(ICartService cartService)
+        private readonly IResourceAuthorizationService _resourceAuthorizationService;
+        public CartController(ICartService cartService, IResourceAuthorizationService resourceAuthorizationService)
         {
             _cartService = cartService;
+            _resourceAuthorizationService = resourceAuthorizationService;
         }
 
         [HttpGet]
@@ -31,6 +32,11 @@ namespace DeliveryDeck_Backend_Final.Controllers
         [ClaimPermissionRequirement(CartPermissions.Adjust)]
         public async Task<IActionResult> AddDish(Guid dishId, [FromQuery, BindRequired] int amount = 1)
         {
+            if (! await _resourceAuthorizationService.DishResourceExists(dishId))
+            {
+                return NotFound();
+            }
+
             await _cartService.AddDish(UserId, dishId, amount);
             return NoContent();
         }
@@ -39,6 +45,14 @@ namespace DeliveryDeck_Backend_Final.Controllers
         [ClaimPermissionRequirement(CartPermissions.Adjust)]
         public async Task<IActionResult> RemoveDish(Guid dishId, [FromQuery, BindRequired] int amount = 1)
         {
+            if (
+                !(await _resourceAuthorizationService.DishResourceExists(dishId)
+                && await _resourceAuthorizationService.DishInCartResourceExists(UserId, dishId))
+                )
+            {
+                return NotFound();
+            }
+
             await _cartService.RemoveDish(UserId, dishId, amount);
             return NoContent();
         }
@@ -47,6 +61,14 @@ namespace DeliveryDeck_Backend_Final.Controllers
         [ClaimPermissionRequirement(CartPermissions.Adjust)]
         public async Task<IActionResult> RemoveDishCompletely(Guid dishId)
         {
+            if (
+                !(await _resourceAuthorizationService.DishResourceExists(dishId)
+                && await _resourceAuthorizationService.DishInCartResourceExists(UserId, dishId))
+                )
+            {
+                return NotFound();
+            }
+
             await _cartService.RemoveDishCompletely(UserId, dishId);
             return NoContent();
         }
